types: fix u64 and s64 type definition
The C standard guarantees 'unsigned long long' to be 64 bits, while
'unsigned long' may be only 32 bits. This shouldn't make any difference in
our case, with the exception of GCC's type checker now being happy without
an explicit cast to (unsigned long long) whenever %llu or %llx is used.

+++ include/types.h
@@ -3,8 +3,8 @@
 
 #define NULL	((void*) 0)
 
-typedef unsigned long u64;
-typedef signed long s64;
+typedef unsigned long long u64;
+typedef signed long long s64;
 
 typedef unsigned int u32;
 typedef signed int s32;
